backend/app/app/tests/api/api_v1/test_login.py
from typing import Dict

from fastapi.testclient import TestClient

from app.core.config import settings


def test_get_access_token(client: TestClient) -> None:
    login_data = {
        "creatorname": settings.FIRST_ADMIN,
        "password": settings.FIRST_ADMIN_PASSWORD,
    }
    r = client.post(f"{settings.API_V1_STR}/login/access-token", data=login_data)
    tokens = r.json()
    assert r.status_code == 200
    assert "access_token" in tokens
    assert tokens["access_token"]


def test_use_access_token(client: TestClient, _token_headers: Dict[str, str]) -> None:
    r = client.post(
        f"{settings.API_V1_STR}/login/test-token",
        headers=_token_headers,
    )
    result = r.json()
    assert r.status_code == 200
    assert "email" in result
